ContinousEval1-main/ContinousEval1-main/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Cors==3.0.10
Flask-Bcrypt==0.7.1
Flask-Migrate==3.1.0
PyJWT==2.1.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
python-dotenv==0.19.0
requests==2.26.0 